+++ include/simple-mpc/kinodynamics.hpp
@@ -7,7 +7,7 @@
 ///////////////////////////////////////////////////////////////////////////////
 #pragma once
 
-#include "simple-mpc/base-problem.hpp"
+#include "simple-mpc/ocp-handler.hpp"
 
 namespace simple_mpc {
 using namespace aligator;
